Parte do Admin praticamente acabada apenas falta ativar e desativar empresas

+++ Trabalho_PWEB/Controllers/EmpresasController.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
@@ -11,11 +12,12 @@ using Trabalho_PWEB.Models;
 
 namespace Trabalho_PWEB.Controllers
 {
+    [Authorize(Roles = "Admin")]
     public class EmpresasController : Controller
     {
         private readonly ApplicationDbContext _context;
         private readonly UserManager<ApplicationUser> _userManager;
-
+        
         public EmpresasController(ApplicationDbContext context,UserManager<ApplicationUser> userManager)
         {
             _context = context;
@@ -27,25 +29,28 @@ namespace Trabalho_PWEB.Controllers
         {
             
             if (opcao == null) {
-                System.Diagnostics.Debug.WriteLine();
                 return View(await _context.Empresa.ToListAsync());
             }else{
                 switch (opcao)
                 {
                     case 1:
-                        System.Diagnostics.Debug.WriteLine("CASE 1");
                         return View(_context.Empresa.Where(e => e.EstadoSubscricao == false).ToList());
                     case 2:
-                        System.Diagnostics.Debug.WriteLine("CASE 2");
                         return View(_context.Empresa.Where(e => e.EstadoSubscricao == true).ToList());
-                    case 3:
-                        break;
                     default:
                         return View();
                         break;    
                 }
             }
             return View();
+        }
+
+        public async Task<IActionResult> Search(string? nome)
+        {
+            if(nome == null || nome == ""){
+                return View("Index",await _context.Empresa.ToListAsync());
+            }
+            return View("Index", _context.Empresa.Where(e => e.Nome.Contains(nome)).ToList());
         }
 
         // GET: Empresas/Details/5
@@ -62,7 +67,23 @@ namespace Trabalho_PWEB.Controllers
             {
                 return NotFound();
             }
-
+            List<ApplicationUser> l = new List<ApplicationUser>();
+            l = _context.Users.Where(u => u.EmpresaId == id).ToList();
+            List<string> rolesUsers = new List<string>();
+            foreach (var user in l)
+            {
+                IEnumerable<string> Roles = await _userManager.GetRolesAsync(_userManager.Users.Where(u => u.Id == user.Id).First());
+                string aux = "";
+                foreach(var item in Roles)
+                {
+                    aux += item + "; ";
+                }
+                rolesUsers.Add(aux);
+            }
+            ViewBag.rolesUsers = rolesUsers;
+            List<Veiculo> v = new List<Veiculo>();
+            v = _context.Veiculo.Where(v => v.EmpresaId == id).ToList();
+            ViewBag.Veiculo = v;
             return View(empresa);
         }
 
@@ -79,10 +100,6 @@ namespace Trabalho_PWEB.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create(CreateEmpresaModelView model)
         {
-            System.Diagnostics.Debug.WriteLine(model.Nome);
-            System.Diagnostics.Debug.WriteLine(model.Email);
-            System.Diagnostics.Debug.WriteLine(model.Password);
-            System.Diagnostics.Debug.WriteLine(model.ConfirmPassword);
             var defaultUser = new ApplicationUser
             {
                 UserName = model.Email,
@@ -95,11 +112,18 @@ namespace Trabalho_PWEB.Controllers
             {
                 await _userManager.CreateAsync(defaultUser, model.Password);
                 await _userManager.AddToRoleAsync(defaultUser, Roles.Gestor.ToString());
+            }else{
+                return View();
             }
             Empresa empresa = new Empresa();
             empresa.Nome = model.Nome;
             empresa.Avaliação = 0;
             empresa.EstadoSubscricao = true;
+            ICollection<ApplicationUser> l= new List<ApplicationUser>();
+            ICollection<Veiculo> v = new List<Veiculo>();
+            l.Add(await _userManager.FindByEmailAsync(defaultUser.Email));
+            empresa.ListaFuncionarios = l;
+            empresa.Veiculos = v;
             _context.Add(empresa);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
@@ -128,12 +152,12 @@ namespace Trabalho_PWEB.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("Id,Nome,EstadoSubscricao,Avaliação")] Empresa empresa)
         {
+
             if (id != empresa.Id)
             {
                 return NotFound();
             }
-
-            if (ModelState.IsValid)
+            if (ModelState.ErrorCount<=2)
             {
                 try
                 {
@@ -186,7 +210,12 @@ namespace Trabalho_PWEB.Controllers
             var empresa = await _context.Empresa.FindAsync(id);
             if (empresa != null)
             {
+                if(_context.Veiculo.Where(v => v.EmpresaId == id).Count() == 0) {
+                _context.Users.RemoveRange(_context.Users.Where(u => u.EmpresaId == id).ToList());
                 _context.Empresa.Remove(empresa);
+                }else{
+                    return Problem("Empresa tem veiculos registados");
+                }
             }
             
             await _context.SaveChangesAsync();
